+++ front/src/hooks/useCollapsableHeader.ts
@@ -1,0 +1,23 @@
+import { RefObject, useEffect, useState } from 'react';
+
+export function useCollapsibleHeader(ref: RefObject<HTMLDivElement | null>) {
+  const [isOpen, setIsOpen] = useState(false);
+  const toggleIsOpen = () => setIsOpen(!isOpen);
+  const closeHeader = () => setIsOpen(false);
+
+  useEffect(() => {
+    function closeHeaderOnOutsideClick(event: Event) {
+      const isClickOutside =
+        ref.current && !ref.current.contains(event.target as Node);
+      if (isClickOutside && isOpen) {
+        closeHeader();
+      }
+    }
+    document.addEventListener('mousedown', closeHeaderOnOutsideClick);
+    return () => {
+      document.removeEventListener('mousedown', closeHeaderOnOutsideClick);
+    };
+  }, [isOpen, ref]);
+
+  return { isOpen, closeHeader, toggleIsOpen };
+}
